after review - transfer two methods to ui
